requirements.txt
BabelDjango==0.2.2
Django==1.5.4
PyYAML==3.10
South==0.7.6
dj-database-url==0.2.1
django-password-reset==0.4
gunicorn==0.17.4
pyaml==13.01.0
python-dateutil==2.1
ipython==0.13.1
psycopg2==2.5 #sudo apt-get install libpq-dev python-dev
django-tastypie==0.9.15
dateutils==0.6.6
static==0.4
dj-static==0.0.5

django-coverage==1.2.4
mock==1.0.1
